requirements.txt
torch
faiss-cpu
pandas
numpy
scikit-learn
fastapi
uvicorn
jinja2
python-multipart
sentence-transformers
wandb
